:sparkles: implement support for i18n in error messages, versioning for tokenex and custom submit handlers
