+++ src/types/types.ts
@@ -2,14 +2,24 @@ type FieldChangeProps = {
 	fieldId: string;
 	element: HTMLElement;
 	error: Record<string, string>;
+	value: string;
 	isValid: boolean;
 };
 
+type i18nMessagesTypes =
+	| 'requiredField'
+	| 'invalidCardNumber'
+	| 'invalidExpirationDate'
+	| 'invalidCVV';
+
 export type CardCollectProps = {
 	displayErrors?: boolean;
-	onFieldChange?: ({ fieldId, element, error, isValid }: FieldChangeProps) => void;
+	onFieldChange?: ({ fieldId, element, error, value, isValid }: FieldChangeProps) => void;
 	validateOnChange?: boolean;
 	displayHelpIcons?: boolean;
+	handleCardValuesOnSubmit?: boolean;
+	i18nMessages?: Record<i18nMessagesTypes, string>;
+	version?: number;
 };
 
 export type SubmitBody = {
@@ -45,6 +55,7 @@ export type ValidateFieldsProps = {
 	cardValue?: string;
 	dateValue?: string;
 	cvvValue?: string;
+	i18nMessages?: Record<i18nMessagesTypes, string>;
 };
 
 export type ErrorData = {
